+++ src/app/core/decorators/auto-unsubscribe.decorator.ts
@@ -1,0 +1,15 @@
+export function AutoUnsubscribe(subName: string = 'sub') {
+  return (constructor: any) => {
+    const original = constructor.prototype.ngOnDestroy;
+
+    constructor.prototype.ngOnDestroy = function () {
+      const sub = this[subName];
+
+      sub?.unsubscribe();
+      
+      if (original && (typeof original === 'function')) {
+        original.apply(this, arguments);
+      }
+    };
+  };
+}