+++ src/app/core/interceptors/timing.interceptor.ts
@@ -1,0 +1,22 @@
+import { Injectable } from '@angular/core';
+import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest, HttpParams } from '@angular/common/http';
+
+import { Observable } from 'rxjs';
+
+@Injectable()
+export class TimingInterceptor implements HttpInterceptor {
+  intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+    let clonedRequest;
+    if (req.url.includes('products')) {
+      clonedRequest = req.clone({
+        params: new HttpParams()
+          .set('timing_interceptor', Date.now().toString())
+      });
+      console.log(clonedRequest);
+    } else {
+      clonedRequest = req;
+    }
+
+    return next.handle(clonedRequest);
+  }
+}