+++ src/app/orders/models/order.model.ts
@@ -1,0 +1,8 @@
+export class OrderModel {
+	constructor(
+		public id: number = null,
+		public firstname: string = '',
+		public lastname: string = '',
+    public tel: string = ''
+	) {}
+}