+++ src/app/products/components/product-form/product-form.component.ts
@@ -2,28 +2,27 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, ParamMap, Router } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
 
-import { ProductModel } from '../../models/product.model';
+import { ProductModel } from './../../models/product.model';
 import { ProductsPromiseService } from '../../services';
 
 @Component({
-  selector: 'app-product-view',
-  templateUrl: './product-view.component.html',
-  styleUrls: ['./product-view.component.scss']
+  templateUrl: './product-form.component.html',
+  styleUrls: ['./product-form.component.scss']
 })
-export class ProductViewComponent implements OnInit {
-
-  products: Promise<Array<ProductModel>>;
+export class ProductFormComponent implements OnInit {
   product: ProductModel;
 
   constructor(
-    private route: ActivatedRoute,
+    private productsPromiseService: ProductsPromiseService,
     private router: Router,
-    private productsPromiseService: ProductsPromiseService) { }
+    private route: ActivatedRoute
+  ) {}
 
   ngOnInit(): void {
     this.product = new ProductModel();
+
     const observer = {
-      next: (task: ProductModel) => (this.product = { ...task }),
+      next: (product: ProductModel) => (this.product = { ...product }),
       error: (err: any) => console.log(err)
     };
     this.route.paramMap
@@ -35,7 +34,16 @@ export class ProductViewComponent implements OnInit {
         }))  
       .subscribe(observer);
   }
-  
+
+  onSaveProduct() {
+    const product = { ...this.product } as ProductModel;
+    const method = product.id ? 'updateProduct' : 'createProduct';
+    console.log(product);
+    this.productsPromiseService[method](product)
+      .then(() => this.onGoBack())
+      .catch(err => console.log(err));
+  }
+
   onGoBack(): void {
     this.router.navigate(['./../../'], { relativeTo: this.route});
   }