+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,11 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Observable } from 'rxjs';
 
 import { CartProductsService } from 'src/app/cart/services/cart-products.service';
 import { UserRoleService } from 'src/app/core/services/user-role.service';
 import { ProductModel } from '../../models/product.model';
-import { ProductsService } from '../../services/products.service';
+import { ProductsPromiseService } from './../../services';
 
 @Component({
   selector: 'app-product-list',
@@ -14,13 +13,13 @@ import { ProductsService } from '../../services/products.service';
 })
 export class ProductListComponent implements OnInit {
 
-  products$: Observable<ProductModel[]>;
+  products: Promise<Array<ProductModel>>;
 
   constructor(
     public userRoleService: UserRoleService,
     private router: Router, 
-    private productsService: ProductsService, 
-    private cartProductsService: CartProductsService
+    private cartProductsService: CartProductsService,
+    private productsPromiseService: ProductsPromiseService
   ) { }
 
   ngOnInit(): void {
@@ -28,18 +27,33 @@ export class ProductListComponent implements OnInit {
   }
 
   getProducts(): void {
-    this.products$ = this.productsService.getProducts();
+    this.products = this.productsPromiseService.getProducts();
   }
 
-  addProductToCart(id: number): void {
-    this.products$.subscribe(val => {
-      const product = val.find(x => x.id === id);
-      this.cartProductsService.addProduct(product);
-    })
+  async addProductToCart(id: number): Promise<void> {
+    const product = (await this.products).find(x => x.id === id);
+    this.cartProductsService.addProduct(product);
   }
 
   onShowProduct(product: ProductModel): void {
     const link = ['/product', product.id];
     this.router.navigate(link);
+  }
+
+  onCreateProduct() {
+    const link = ['/product/add'];
+    this.router.navigate(link);
+  }
+
+  onEditProduct(product: ProductModel) {
+    const link = ['/edit', product.id];
+    this.router.navigate(link);
+  }
+
+  onDeleteProduct(product: ProductModel): void {
+    this.productsPromiseService
+      .deleteProduct(product)
+      .then(() => (this.products = this.productsPromiseService.getProducts()))
+      .catch(err => console.log(err));
   }
 }