+++ src/app/products/models/product.model.ts
@@ -1,10 +1,11 @@
 import { Сategory } from './category.enum';
-
-export interface ProductModel {
-  id: number;
-  name: string;
-  description: string;
-  price: number;
-  category: Сategory;
-  isAvailable: boolean;
+export class ProductModel {
+  constructor(
+    public id: number = null,
+    public name: string = '',
+    public description: string = '',
+    public price: number = 0,
+    public category?: Сategory,
+    public isAvailable?: boolean
+  ) {}
 }